size, capacity, next, prev, free_elem from int to  size_t

+++ Include/list.h
@@ -11,7 +11,7 @@
 #define INTERNAL_DUMP(list, error)  show_dump(list, error,    dump_position {__FILE__, __FUNCTION__,  __LINE__})
 
 static const int    kStartListCapacity  = 16;
-static const int    kCanary             = 0xDEADBEEF; // TODO implicit cast (unsigned to signed)
+static const int    kCanary             = (int) 0xDEADBEEF;
 static const int    kImageSize          = 1600;
 static const int    kMaxFileNameLen     = 16;
 static const size_t kNoPtr              = SIZE_MAX;
